Fixed bug where points outside Polygon where added to selection
The problem lied in that  if the function utilized integers, there were going to be points that when the line drawn from that point to Infinity would intersect the polygon in the exact corner of the vertex. One way could be to specifically fix that condition, but the variation of double calculation can practically eliminate the posibility that the drawn line to infinity will intersect exactly at a corner point of the polygon..... (UGLY BUG !!! 😣)

+++ include/Polygon.h
@@ -6,11 +6,11 @@
 // Code to check if a given point lies inside a given polygon 
 // Based on https://www.geeksforgeeks.org/check-if-two-given-line-segments-intersect/ 
 namespace GUI {
-	struct Point { int x; int y; };
+	struct Point { double x; double y; };
 	class Polygon {
 	public:
-		bool isInside(int x, int y);
-		inline void addVertex(int x, int y) { m_vertices.push_back(Point{ x,y }); };
+		bool isInside(double x, double y);
+		inline void addVertex(double x, double y) { m_vertices.push_back(Point{ x,y }); };
 		inline void clearVertices() { m_vertices.clear(); };
 	private:
 		std::vector<Point> m_vertices;
